requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.5
boto3==1.17.49
botocore==1.20.49
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
channels==2.4.0
channels-redis==3.0.1
colorama==0.4.4
cryptography==3.4.6
daphne==2.5.0
defusedxml==0.7.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-absoluteuri==1.3.0
django-allauth==0.44.0
django-countable-field==1.3
django-crispy-forms==1.11.1
django-heroku==0.3.1
django-storages==1.11.1
filelock==3.0.12
flake8==3.9.0
gunicorn==20.0.4
idna==2.10
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.1.2
pipenv==2020.11.15
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.0
PyJWT==2.0.1
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==2.10.5
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.6
six==1.15.0
spacy>=2.2.0,<3.0.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.3
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm